requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-chroma
chromadb
pypdf